Fixed bug where the `np` package had to be loaded manually
The R package `np` has some internal functions that are not properly declared as `np::FUN`, so it doesn't work without actively loading the `np` package into the work path. This was fixed by moving the `PRE` package dependencies to the attribute `Depends` instead of `Imports`. This means that the packages are automatically loaded whenever the `PRE` package is loaded. This fixes the bug.